Fixed Converter.java

+++ src/gui/InfoGUI.java
@@ -11,9 +11,9 @@ public class InfoGUI {
             graphics2D.setColor(Color.BLACK);
             graphics2D.setFont(new Font("Arial", Font.PLAIN, 24));
             FontMetrics fontMetrics = graphics2D.getFontMetrics();
-            String coordinate_str = "x: " + camera.x + ", y: " + camera.y + ", z: " + camera.z;
+            String coordinate_str = "x: " + (int) camera.x + ", y: " + (int) camera.y + ", z: " + (int) camera.z;
             graphics2D.drawString(coordinate_str, 0, fontMetrics.getAscent());
-            String rotation_str = "yaw: " + camera.yaw + ", pitch: " + camera.pitch + ", roll: " + camera.roll;
+            String rotation_str = "yaw: " + (int) camera.yaw + ", pitch: " + (int) camera.pitch + ", roll: " + (int) camera.roll;
             graphics2D.drawString(rotation_str, 0, fontMetrics.getHeight() + fontMetrics.getAscent());
             String focal_length_str = "Focal Length: " + camera.focal_length;
             graphics2D.drawString(focal_length_str, 0, 2*fontMetrics.getHeight() + fontMetrics.getAscent());
